adding tests for api

+++ backend/app.py
@@ -16,11 +16,15 @@ from dotenv import load_dotenv
 import os 
 
 
-def createApp(): 
+def createApp(testing = False): 
 
     app = Flask(__name__)
 
     app.config.from_object(Config)
+
+    if testing: 
+        app.config['TESTING'] = True 
+        app.config["SQLALCHEMY_DATABASE_URI"] = app.config["TESTING_SQLALCHEMY_DATABASE_URI"]
 
     load_dotenv()
     app.config["SECRET_KEY"] = os.getenv("SECRET_KEY")
@@ -49,7 +53,7 @@ def createApp():
     @app.route('/dashboard', methods=['GET', 'POST'])
     @login_required
     def dashboard(): 
-        return "this is your dashboard"
+        return jsonify({"msg": "this is your dashboard"}), 200
 
     # The route() function of the Flask class is a decorator, 
     # which tells the application which URL should call 
@@ -65,7 +69,7 @@ def createApp():
             "difficulty": difficulty, 
             "type": type, 
             "question": response
-        })
+        }), 200
 
     @app.route('/generate_question', methods=['POST'])
     def generateQuestion():
@@ -79,65 +83,73 @@ def createApp():
             "difficulty": difficulty, 
             "type": type, 
             "question": response
-        })
+        }), 200
 
     @app.route('/login', methods=['POST'])
     def login():
         data = request.get_json()
         username = data.get('username')
         password = data.get('password')
+        if(username == None or password == None): 
+            return jsonify({"msg": "Username or password cannot be missing"}), 400
 
         user = User.query.filter_by(username=username).first()
         if user: 
             if user.checkPassword(password):
                 login_user(user)
-                return redirect(url_for('dashboard'))
+                return jsonify({"msg": "Logged in user"}), 200
             else:
-                return "Wrong password", 400
+                return jsonify({"msg": "Wrong password"}), 400
         
-        return "No user with that username exists", 404
+        return jsonify({"msg": "No user with that username exists"}), 404
 
     @app.route('/logout', methods=['POST'])
     @login_required
     def logout():
         logout_user()
-        return redirect(url_for('login'))
-    
+        return {"msg": "Logged out user"}, 200   
         
     @app.route('/register', methods=['POST'])
     def register():
         data = request.get_json()
         username = data.get('username')
         password = data.get('password')
-        hashedPassword = bcrypt.generate_password_hash(password)
-        user = User(username=username, password=hashedPassword)
+        if(username == None or password == None): 
+            return jsonify({"msg": "Username or password cannot be missing"}), 400
+        existingUser = User.query.filter_by(username=username).first()
+        if existingUser: 
+            return jsonify({"msg": "User with that username already exists"}), 400
+        user = User(username=username, plainTextPassword=password)
         db.session.add(user)
         db.session.commit()
-        return redirect(url_for('dashboard'))
-
+        return jsonify({"msg": "Registered user"}), 200
 
     @app.route('/users', methods=['GET'])
     def getUsers():
         users = User.query.all()
         if len(users) == 0: 
-            return "User not found", 404
+            return jsonify({"msg": "No users found"}), 404
         return jsonify(users), 200
 
     @app.route('/users/<username>', methods=['GET'])
     def getUser(username): 
         user = User.query.filter_by(username=username).first()
         if not user: 
-            return "User not found", 404
+            return jsonify({"msg": "User not found"}), 404
         return jsonify(user), 200
 
     @app.route('/users/<username>', methods=['PATCH'])
     def updateUser(username): 
-        user = User.query.filter_by(username=username).first()
-        if not user: 
-            return "User not found", 404
         data = request.get_json()
         newUsername = data.get("username")
         newPassword = data.get("password")
+        if(newUsername == None or newPassword == None): 
+            return jsonify({"msg": "Username or password cannot be missing"}), 400
+        
+        user = User.query.filter_by(username=username).first()
+        if not user: 
+            return jsonify({"msg": "User not found"}), 404
+        
         user.username = newUsername
         user.password = newPassword
         db.session.commit()
@@ -145,9 +157,13 @@ def createApp():
 
     @app.route('/users/<username>', methods=['DELETE'])
     def deleteUser(username): 
+        if(username == None): 
+            return jsonify({"msg": "Username cannot be missing"}), 400
+        
         user = User.query.filter_by(username=username).first()
         if not user: 
-            return "User not found", 404
+            return jsonify({"msg": "User not found"}), 404
+        
         db.session.delete(user)
         db.session.commit()
         newUsers = User.query.all()
@@ -158,13 +174,13 @@ def createApp():
     def saveQuestion(username): 
         data = request.get_json()
         question = SavedQuestion(data.get("category"), 
-                                    data.get("difficulty"), 
-                                    data.get("type"), 
-                                    data.get("question"), 
-                                    data.get("answer"), 
-                                    data.get("userAnswer"), 
-                                    data.get("notes"), 
-                                    data.get("userId"))
+                                data.get("difficulty"), 
+                                data.get("type"), 
+                                data.get("question"), 
+                                data.get("answer"), 
+                                data.get("userAnswer"), 
+                                data.get("notes"), 
+                                data.get("userId"))
         db.session.add(question)
         db.session.commit()
         return jsonify(question), 201
@@ -175,7 +191,7 @@ def createApp():
         data = request.get_json()
         question = SavedQuestion.query.filter_by(userId=username, id=questionId).first()
         if not question: 
-            return "Question not found", 404
+            return jsonify({"msg": f"Question for {username} was not found"}), 404
         return jsonify(question), 200
     
     return app
